Rename keyvaultclient to keyvault
