+++ pkg/certificate/providers/keyvault/client.go
@@ -1,4 +1,4 @@
-package keyvaultclient
+package keyvault
 
 import (
 	"fmt"
