+++ pkg/certificate/providers/keyvault/types.go
@@ -1,4 +1,4 @@
-package keyvaultclient
+package keyvault
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/keyvault/v7.0/keyvault"
